distsqlrun: move mocking test utils to their own file
In the next commit I'll add to this file and use it from a different
test file in the distsqlrun package.
